refactor : lotto 상위 packaging
