+++ docs/lottogame/step_two_feature_list.md
@@ -1,7 +1,7 @@
 # 프로젝트 전체
-- [ ] : lotto 상위 package 추가... 
-- [ ] : 한 메서드에 오직 한 단계의 들여쓰기만 한다.
-- [ ] : 함수의 길이가 15라인을 넘어가지 않도록 구현한다
+- [x] : lotto 상위 package 추가... 
+- [x] : 한 메서드에 오직 한 단계의 들여쓰기만 한다.
+- [x] : 함수의 길이가 15라인을 넘어가지 않도록 구현한다
 ## View
 - [x] : 수동으로 구매할 로또 수 입력 받기(input) - n
 - [x] : 수동으로 구매할 번호들을 입력(input) - List<List<Int>> 반환
