+++ src/main/kotlin/lottogame/model/LottoNumber.kt
@@ -1,0 +1,5 @@
+package lottogame.model
+
+interface LottoNumber {
+    val number: Int
+}
