+++ src/main/kotlin/lottogame/view/LottoGameInputView.kt
@@ -1,0 +1,13 @@
+package lottogame.view
+
+interface LottoGameInputView {
+    fun inputPurchaseExpense(): Int
+
+    fun inputManualLottoCount(): Int
+
+    fun inputManualLottoNumbers(size: Int): List<List<Int>>
+
+    fun inputWinningNumbers(): List<Int>
+
+    fun inputBonusNumber(): Int
+}
