--- src/main/kotlin/model/LottoNumber.kt
@@ -1,5 +1,0 @@
-package model
-
-interface LottoNumber {
-    val number: Int
-}
