--- src/main/kotlin/model/ManualLottieGenerator.kt
@@ -1,11 +1,0 @@
-package model
-
-class DefaultManualLottieGenerator() : ManualLottieGenerator {
-    override fun generate(lottieNumbers: List<List<Int>>): List<Lotto> {
-        return lottieNumbers.map { Lotto(it) }
-    }
-}
-
-fun interface ManualLottieGenerator {
-    fun generate(lottieNumbers: List<List<Int>>): List<Lotto>
-}
