--- src/main/kotlin/view/LottoGameInputView.kt
@@ -1,13 +1,0 @@
-package view
-
-interface LottoGameInputView {
-    fun inputPurchaseExpense(): Int
-
-    fun inputManualLottoCount(): Int
-
-    fun inputManualLottoNumbers(size: Int): List<List<Int>>
-
-    fun inputWinningNumbers(): List<Int>
-
-    fun inputBonusNumber(): Int
-}
